Adding baseline functionality for assertions upon an executable's output/error, wall clock run time, memory consumption, processor time, or exit code
There's a problem where you cannot use Process to access both a sequentially consistent output (as you might see in a console application) and a separated output at the same time; it's not even clear if Windows supports doing this whatsoever. This informed the decision to create 3 classes where one is just an interface to the other 2 - this means that the process is executed twice and has performance implications. The API is set up to choose which sort of output you want at the class level. I still need to add some mechanism for either supporting input into processes that are requesting input or detecting that input is required, but this is a good start.